requirements.txt
numpy~=1.24.3
pyproj~=3.5.0
rasterio~=1.3.6
shapely~=2.0.1
requests~=2.29.0
expression-parser==0.0.5